+++ common/src/main/java/com/bugzero/rarego/shared/payment/dto/SettlementResponseDto.java
@@ -1,0 +1,15 @@
+package com.bugzero.rarego.shared.payment.dto;
+
+import java.time.LocalDateTime;
+
+public record SettlementResponseDto(
+	Long id,
+	Long auctionId,
+	Long sellerId,
+	int salesAmount,
+	int feeAmount,
+	int settlementAmount,
+	String status,
+	LocalDateTime createdAt
+) {
+}
