--- payment-service/src/main/java/com/bugzero/rarego/event/SettlementFinishedEvent.java
@@ -1,4 +1,0 @@
-package com.bugzero.rarego.event;
-
-public record SettlementFinishedEvent() {
-}
