+++ payment-service/src/main/java/com/bugzero/rarego/in/PaymentEventListener.java
@@ -10,7 +10,7 @@ import org.springframework.transaction.event.TransactionalEventListener;
 
 import com.bugzero.rarego.app.PaymentFacade;
 import com.bugzero.rarego.app.PaymentSettlementProcessor;
-import com.bugzero.rarego.event.SettlementFinishedEvent;
+import com.bugzero.rarego.shared.payment.event.SettlementFinishedEvent;
 import com.bugzero.rarego.shared.auction.event.AuctionEndedEvent;
 import com.bugzero.rarego.shared.member.event.MemberJoinedEvent;
 import com.bugzero.rarego.shared.member.event.MemberUpdatedEvent;
