--- payment-service/src/main/java/com/bugzero/rarego/in/dto/SettlementResponseDto.java
@@ -1,30 +1,0 @@
-package com.bugzero.rarego.in.dto;
-
-import java.time.LocalDateTime;
-
-import com.bugzero.rarego.domain.Settlement;
-import com.bugzero.rarego.domain.SettlementStatus;
-
-public record SettlementResponseDto(
-	Long id,
-	Long auctionId,
-	Long sellerId,
-	int salesAmount,
-	int feeAmount,
-	int settlementAmount,
-	SettlementStatus status,
-	LocalDateTime createdAt
-) {
-	public static SettlementResponseDto from(Settlement settlement) {
-		return new SettlementResponseDto(
-			settlement.getId(),
-			settlement.getAuctionId(),
-			settlement.getSeller().getId(),
-			settlement.getSalesAmount(),
-			settlement.getFeeAmount(),
-			settlement.getSettlementAmount(),
-			settlement.getStatus(),
-			settlement.getCreatedAt()
-		);
-	}
-}
