+++ payment-service/src/test/java/com/bugzero/rarego/app/PaymentProcessSettlementUseCaseUnitTest.java
@@ -14,11 +14,9 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import com.bugzero.rarego.app.PaymentProcessSettlementUseCase;
-import com.bugzero.rarego.app.PaymentSettlementProcessor;
 import com.bugzero.rarego.domain.Settlement;
 import com.bugzero.rarego.domain.SettlementStatus;
-import com.bugzero.rarego.event.SettlementFinishedEvent;
+import com.bugzero.rarego.shared.payment.event.SettlementFinishedEvent;
 import com.bugzero.rarego.out.SettlementRepository;
 import com.bugzero.rarego.global.event.EventPublisher;
