+++ payment-service/src/test/java/com/bugzero/rarego/in/PaymentControllerTest.java
@@ -26,14 +26,13 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
 import com.bugzero.rarego.app.PaymentFacade;
-import com.bugzero.rarego.in.PaymentController;
 import com.bugzero.rarego.in.dto.AuctionFinalPaymentRequestDto;
 import com.bugzero.rarego.in.dto.AuctionFinalPaymentResponseDto;
 import com.bugzero.rarego.in.dto.PaymentConfirmRequestDto;
 import com.bugzero.rarego.in.dto.PaymentConfirmResponseDto;
 import com.bugzero.rarego.in.dto.PaymentRequestDto;
 import com.bugzero.rarego.in.dto.PaymentRequestResponseDto;
-import com.bugzero.rarego.in.dto.SettlementResponseDto;
+import com.bugzero.rarego.shared.payment.dto.SettlementResponseDto;
 import com.bugzero.rarego.in.dto.WalletResponseDto;
 import com.bugzero.rarego.in.dto.WalletTransactionResponseDto;
 import com.bugzero.rarego.domain.SettlementStatus;
